moved all user authentication logic to a separate file
